# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RelancePro Africa API",
    "description": "API publique pour la gestion des créances clients en Afrique. Support REST et GraphQL.",
    "version": "1.0.0",
    "contact": {
      "name": "RelancePro Africa Support",
      "email": "support@relancepro.africa"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.relancepro.africa",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Utilisez votre clé API comme Bearer token. Exemple: Authorization: Bearer rpa_live_xxxxx"
      }
    },
    "schemas": {
...
      "Reminder": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string", "enum": ["email", "whatsapp"] },
          "subject": { "type": "string", "nullable": true },
          "message": { "type": "string" },
          "tone": { "type": "string", "enum": ["formal", "friendly", "urgent"], "nullable": true },
          "status": { "type": "string", "enum": ["pending", "sent", "failed", "delivered", "opened"] },
          "sentAt": { "type": "string", "format": "date-time", "nullable": true },
          "deliveredAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Webhook": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "url": { "type": "string" },
          "events": { "type": "array", "items": { "type": "string" } },
          "active": { "type": "boolean" },
          "lastTriggeredAt": { "type": "string", "format": "date-time", "nullable": true },
          "successCount": { "type": "integer" },
          "failureCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object", "nullable": true }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "perPage": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "hasMore": { "type": "boolean" }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/reminders": {
      "get": {
        "tags": ["Reminders"],
        "summary": "List reminders",
        "operationId": "listReminders",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "perPage", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "sent", "failed", "delivered", "opened"] } },
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["email", "whatsapp"] } },
          { "name": "clientId", "in": "query", "schema": { "type": "string" } },
          { "name": "debtId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Reminder" } },
                    "meta": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" }
        }
      },
      "post": {
        "tags": ["Reminders"],
        "summary": "Create and send reminder",
        "operationId": "createReminder",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["debtId", "type"],
                "properties": {
                  "debtId": { "type": "string" },
                  "type": { "type": "string", "enum": ["email", "whatsapp"] },
                  "subject": { "type": "string" },
                  "message": { "type": "string" },
                  "tone": { "type": "string", "enum": ["formal", "friendly", "urgent"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reminder created and sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Reminder" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "description": "Debt not found" },
          "500": { "description": "Failed to send reminder" }
        }
      }
    },
    "/api/v1/webhooks": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "List webhooks",
        "operationId": "listWebhooks",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "perPage", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Webhook" } },
                    "meta": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedError" }
        }
      },
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create webhook",
        "operationId": "createWebhook",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "url", "events"],
                "properties": {
                  "name": { "type": "string" },
                  "url": { "type": "string", "format": "uri" },
                  "events": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "client.created", "client.updated", "client.deleted",
                        "debt.created", "debt.updated", "debt.paid", "debt.deleted",
                        "reminder.sent", "reminder.delivered", "reminder.failed",
                        "payment.received"
                      ]
                    }
                  },
                  "active": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Webhook" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" }
        }
      },
      "delete": {
        "tags": ["Webhooks"],
        "summary": "Delete webhook",
        "operationId": "deleteWebhook",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Webhook deleted" },
          "400": { "description": "Webhook ID required" },
          "401": { "$ref": "#/components/responses/UnauthorizedError" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      }
    },
...
  },
...
}